requirements.txt
amqp==5.3.1
annotated-types==0.7.0
arrow==1.4.0
asgiref==3.11.0
billiard==4.2.4
binaryornot==0.4.4
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
crispy-bootstrap5==2025.6
crispy-tailwind==1.0.3
cryptography==46.0.4
cssbeautifier==1.15.4
cycler==0.12.1
distlib==0.4.0
Django==6.0.1
django-allauth==65.14.0
django-browser-reload==1.21.0
django-crispy-forms==2.5
django-tailwind==4.4.2
django-unfold==0.80.2
django-widget-tweaks==1.5.1
dotenv==0.9.9
EditorConfig==0.17.1
filelock==3.20.0
fonttools==4.60.1
google-ai-generativelanguage==0.6.10
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
httplib2==0.31.2
idna==3.11
Jinja2==3.1.6
jmespath==1.1.0
jsbeautifier==1.15.4
json5==0.13.0
kiwisolver==1.4.9
kombu==5.6.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
pillow==12.1.0
pipenv==2025.0.4
platformdirs==4.5.1
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==5.29.6
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.2.5
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-slugify==8.0.4
pytokens==0.4.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.2.0
s3transfer==0.16.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
virtualenv==20.35.3
wcwidth==0.5.3
